OC-679 ~ Processes review comments

+++ integration-tests/cucumber-tests-platform-distributionautomation/src/test/java/org/opensmartgridplatform/cucumber/platform/distributionautomation/PlatformDistributionAutomationDefaults.java
@@ -15,7 +15,7 @@ public class PlatformDistributionAutomationDefaults {
 
     public static final Integer COMMON_ADDRESS = 1511;
     public static final String FEEDER_NAME = "feeder-1";
-    public static final Long FEEDER_NUMBER = 1L;
+    public static final Integer FEEDER_NUMBER = 1;
     public static final String MQTT_HOST = "0.0.0.0";
     public static final Integer MQTT_PORT = 8883;
     public static final String MQTT_QOS = "EXACTLY_ONCE";
